requirements.txt
asgiref==3.11.0
Django==5.2.8
django-environ==0.12.0
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
# weasyprint and deps optional for PDF; install locally if needed
# weasyprint==66.0
# fonttools==4.53.0
# tinycss2==1.3.0
# cssselect2==0.7.0
# pydyf==0.11.0
# pyphen==0.15.0
